+++ src/ResourceManager/ApiManagement/Commands.ApiManagement.ServiceManagement/Commands/SetAzureRmApiManagementTenantAccess.cs
@@ -17,7 +17,7 @@ namespace Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Commands
     using Microsoft.Azure.Commands.ApiManagement.ServiceManagement.Models;
     using System.Management.Automation;
 
-    [Cmdlet(VerbsCommon.Set, Constants.ApiManagementTenantAccess)]
+    [Cmdlet("Set", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "ApiManagementTenantAccess")]
     [OutputType(typeof(PsApiManagementAccessInformation))]
     public class SetAzureRmApiManagementTenantAccess : AzureApiManagementCmdletBase
     {