+++ src/ResourceManager/DataFactoryV2/Commands.DataFactoryV2/Datasets/SetAzureDataFactoryDatasetCommand.cs
@@ -20,8 +20,7 @@ using Microsoft.WindowsAzure.Commands.Utilities.Common;
 
 namespace Microsoft.Azure.Commands.DataFactoryV2
 {
-    [Cmdlet(VerbsCommon.Set, Constants.Dataset, DefaultParameterSetName = ParameterSetNames.ByFactoryName,
-        SupportsShouldProcess = true), OutputType(typeof(PSDataset))]
+    [Cmdlet("Set", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataFactoryV2Dataset", DefaultParameterSetName = ParameterSetNames.ByFactoryName,SupportsShouldProcess = true), OutputType(typeof(PSDataset))]
     [Alias(VerbsCommon.New + "-" + Constants.Dataset)]
     public class SetAzureDataFactoryDatasetCommand : DataFactoryContextBaseSetCmdlet
     {