+++ src/ResourceManager/Network/Commands.Network/LoadBalancer/LoadBalancerRule/RemoveAzureLoadBalancerRuleCommand.cs
@@ -18,7 +18,7 @@ using System.Management.Automation;
 
 namespace Microsoft.Azure.Commands.Network
 {
-    [Cmdlet(VerbsCommon.Remove, "AzureRmLoadBalancerRuleConfig"), OutputType(typeof(PSLoadBalancer))]
+    [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "LoadBalancerRuleConfig"), OutputType(typeof(PSLoadBalancer))]
     public class RemoveAzureLoadBalancerRuleCommand : NetworkBaseCmdlet
     {
         [Parameter(