+++ src/ResourceManager/Websites/Commands.Websites/Cmdlets/DeploymentSlots/SetAzureWebAppSlotConfigName.cs
@@ -18,7 +18,7 @@ using Microsoft.Azure.Management.WebSites.Models;
 
 namespace Microsoft.Azure.Commands.WebApps.Cmdlets.DeploymentSlots
 {
-    [Cmdlet(VerbsCommon.Set, "AzureRmWebAppSlotConfigName"), OutputType(typeof(SlotConfigNamesResource))]
+    [Cmdlet("Set", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "WebAppSlotConfigName"), OutputType(typeof(SlotConfigNamesResource))]
     public class SetAzureWebAppSlotConfigName : WebAppBaseCmdlet
     {
         [Parameter(Position = 2, Mandatory = false, HelpMessage = "Names of app settings that need to marked as slot settings", ValueFromPipelineByPropertyName = true)]