+++ tools/StaticAnalysis/SignatureVerifier/SignatureVerifier.cs
@@ -132,7 +132,12 @@ namespace StaticAnalysis.SignatureVerifier
                             {
                                 issueLogger.Decorator.AddDecorator(a => a.AssemblyFileName = assemblyFile, "AssemblyFileName");
                                 processedHelpFiles.Add(assemblyFile);
-                                var proxy = EnvironmentHelpers.CreateProxy<CmdletLoader>(directory, out _appDomain);
+                                var proxy =
+#if !NETSTANDARD
+                                    EnvironmentHelpers.CreateProxy<CmdletLoader>(directory, out _appDomain);
+#else
+                                    new CmdletLoader();
+#endif
                                 var module = proxy.GetModuleMetadata(assemblyFile, requiredModules);
                                 var cmdlets = module.Cmdlets;