+++ src/ResourceManager/EventHub/Commands.EventHub/ChangeLog.md
@@ -19,8 +19,12 @@
 -->
 ## Current Release
 
+## Version 0.6.9
+* Fixed issue with default resource groups not being set.
+
 ## Version 0.6.8
 * Updated to the latest version of the Azure ClientRuntime.
+* Updated exception handling for cmdlets
 
 ## Version 0.6.7
 * Updated piping for InputObject and ResourceId in remove cmdlets